+++ main_test.go
@@ -2,15 +2,66 @@ package main
 
 import (
 	"encoding/json"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"net/url"
 	"testing"
 
 	"github.com/opsgenie/opsgenie-go-sdk-v2/alert"
 	"github.com/opsgenie/opsgenie-go-sdk-v2/client"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
+func clearConfig() {
+	plugin.ApiUrl = ""
+	plugin.AuthToken = ""
+}
+
+func TestCreateAlert(t *testing.T) {
+	clearConfig()
+	assert := assert.New(t)
+	event := corev2.FixtureEvent("entity1", "check1")
+	event.Check.Status = 1
+	event.Metrics = nil
+
+	var apiStub = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		body, _ := ioutil.ReadAll(r.Body)
+		expectedBody := `"message":"entity1/check1"`
+		assert.Contains(string(body), expectedBody)
+		w.Header().Add("X-RateLimit-State", "OK")
+		w.WriteHeader(http.StatusAccepted)
+		_, err := w.Write([]byte(`{"result": "Request will be processed",
+                                           "took": 0.302,
+                                           "requestId": "43a29c5c-3dbf-4fa4-9c26-f4f71023e120"
+                                           }`))
+		require.NoError(t, err)
+	}))
+	url, err := url.Parse(apiStub.URL)
+	assert.NoError(err)
+
+	plugin.ApiUrl = url.Host
+	plugin.AuthToken = "test_token"
+
+	plugin.MessageTemplate = "{{.Entity.Name}}/{{.Check.Name}}"
+	plugin.MessageLimit = 200
+	title, _, _ := parseEventKeyTags(event)
+	assert.Equal("entity1/check1", title)
+	alertClient, err := alert.NewClient(&client.Config{
+		ApiKey:         plugin.AuthToken,
+		OpsGenieAPIURL: client.ApiUrl(plugin.ApiUrl),
+	})
+
+	assert.NoError(err)
+
+	err = createIncident(alertClient, event)
+	assert.NoError(err)
+}
+
 func TestGetNote(t *testing.T) {
+	clearConfig()
 	event := corev2.FixtureEvent("foo", "bar")
 	eventJSON, err := json.Marshal(event)
 	assert.NoError(t, err)
@@ -21,6 +72,7 @@ func TestGetNote(t *testing.T) {
 }
 
 func TestParseEventKeyTags(t *testing.T) {
+	clearConfig()
 	event := corev2.FixtureEvent("foo", "bar")
 	_, err := json.Marshal(event)
 	assert.NoError(t, err)
@@ -34,6 +86,7 @@ func TestParseEventKeyTags(t *testing.T) {
 }
 
 func TestParseDescription(t *testing.T) {
+	clearConfig()
 	event := corev2.FixtureEvent("foo", "bar")
 	event.Check.Output = "Check OK"
 	_, err := json.Marshal(event)
@@ -45,6 +98,7 @@ func TestParseDescription(t *testing.T) {
 }
 
 func TestParseDetails(t *testing.T) {
+	clearConfig()
 	event := corev2.FixtureEvent("foo", "bar")
 	event.Check.Output = "Check OK"
 	event.Check.State = "passing"
@@ -58,6 +112,7 @@ func TestParseDetails(t *testing.T) {
 }
 
 func TestEventPriority(t *testing.T) {
+	clearConfig()
 	testcases := []struct {
 		myPriority         string
 		mismatchedPriority string
@@ -81,6 +136,7 @@ func TestEventPriority(t *testing.T) {
 }
 
 func TestCheckArgs(t *testing.T) {
+	clearConfig()
 	assert := assert.New(t)
 	event := corev2.FixtureEvent("entity1", "check1")
 	assert.Error(checkArgs(event))
@@ -91,6 +147,7 @@ func TestCheckArgs(t *testing.T) {
 }
 
 func TestStringInSlice(t *testing.T) {
+	clearConfig()
 	testSlice := []string{"foo", "bar", "test"}
 	testString := "test"
 	testResult := stringInSlice(testString, testSlice)
@@ -101,12 +158,19 @@ func TestStringInSlice(t *testing.T) {
 }
 
 func TestTrim(t *testing.T) {
+	clearConfig()
 	testString := "This string is 33 characters long"
 	assert.Equal(t, trim(testString, 40), testString)
 	assert.Equal(t, trim(testString, 4), "This")
 }
 
 func TestSwitchOpsgenieRegion(t *testing.T) {
+
+	clearConfig()
+	plugin.ApiUrl = "test-host:test-port"
+	testVal := switchOpsgenieRegion()
+	assert.Equal(t, testVal, client.ApiUrl(plugin.ApiUrl))
+	clearConfig()
 	expectedValueUS := client.API_URL
 	expectedValueEU := client.API_URL_EU